Componentize edit and save functions
