+++ src/blocks/block-author-profile/index.js
@@ -2,40 +2,19 @@
  * BLOCK: Atomic Blocks Profile Box
  */
 
-// Import block dependencies and components
-import classnames from 'classnames';
-import Inspector from './components/inspector';
-import ProfileBox from './components/profile';
-import SocialIcons from './components/social';
-import AvatarColumn from './components/avatar';
-import icons from './components/icons';
-
-// Import styles
+/**
+ * Internal dependencies
+ */
+import Edit from './components/edit';
+import Save from './components/save';
 import './styles/style.scss';
 import './styles/editor.scss';
 
-// Internationalization
+/**
+ * WordPress dependencies
+ */
 const { __ } = wp.i18n;
-
-// Extend component
-const { Component } = wp.element;
-
-// Register block
 const { registerBlockType } = wp.blocks;
-
-// Register components
-const {
-	RichText,
-	AlignmentToolbar,
-	BlockControls,
-	InspectorControls,
-	MediaUpload,
-} = wp.editor;
-
-// Register Inspector components
-const {
-	Button,
-} = wp.components;
 
 const blockAttributes = {
 	profileName: {
@@ -117,141 +96,9 @@ const blockAttributes = {
 	},
 };
 
-const ALLOWED_MEDIA_TYPES = [ 'image' ];
-
-class ABAuthorProfileBlock extends Component {
-
-	render() {
-
-		// Setup the attributes
-		const {
-			attributes: {
-				profileName,
-				profileTitle,
-				profileContent,
-				profileAlignment,
-				profileImgURL,
-				profileImgID,
-				profileFontSize,
-				profileBackgroundColor,
-				profileTextColor,
-				profileLinkColor,
-				twitter,
-				facebook,
-				instagram,
-				pinterest,
-				google,
-				youtube,
-				github,
-				email,
-				website,
-				profileAvatarShape
-			},
-			attributes,
-			isSelected,
-			editable,
-			className,
-			setAttributes
-		} = this.props;
-
-		const onSelectImage = img => {
-			setAttributes( {
-				profileImgID: img.id,
-				profileImgURL: img.url,
-			} );
-		};
-
-		return [
-			// Show the block alignment controls on focus
-			<BlockControls key="controls">
-				<AlignmentToolbar
-					value={ profileAlignment }
-					onChange={ ( value ) => setAttributes( { profileAlignment: value } ) }
-				/>
-			</BlockControls>,
-			// Show the block controls on focus
-			<Inspector
-				{ ...{ setAttributes, ...this.props } }
-			/>,
-			// Show the block markup in the editor
-			<ProfileBox { ...this.props }>
-				<AvatarColumn { ...this.props }>
-					<div className="ab-profile-image-square">
-						<MediaUpload
-							buttonProps={ {
-								className: 'change-image'
-							} }
-							onSelect={ ( img ) => setAttributes(
-								{
-									profileImgID: img.id,
-									profileImgURL: img.url,
-								}
-							) }
-							allowed={ ALLOWED_MEDIA_TYPES }
-							type="image"
-							value={ profileImgID }
-							render={ ( { open } ) => (
-								<Button onClick={ open }>
-									{ ! profileImgID ? icons.upload : <img
-										className="profile-avatar"
-										src={ profileImgURL }
-										alt="avatar"
-									/>  }
-								</Button>
-							) }
-						>
-						</MediaUpload>
-					</div>
-				</AvatarColumn>
-
-				<div
-					className={ classnames(
-						'ab-profile-column ab-profile-content-wrap'
-					) }
-				>
-					<RichText
-						tagName="h2"
-						placeholder={ __( 'Add name', 'atomic-blocks' ) }
-						keepPlaceholderOnFocus
-						value={ profileName }
-						className='ab-profile-name'
-						style={ {
-							color: profileTextColor
-						} }
-						onChange={ ( value ) => setAttributes( { profileName: value } ) }
-					/>
-
-					<RichText
-						tagName="p"
-						placeholder={ __( 'Add title', 'atomic-blocks' ) }
-						keepPlaceholderOnFocus
-						value={ profileTitle }
-						className='ab-profile-title'
-						style={ {
-							color: profileTextColor
-						} }
-						onChange={ ( value ) => setAttributes( { profileTitle: value } ) }
-					/>
-
-					<RichText
-						tagName="div"
-						className='ab-profile-text'
-						multiline="p"
-						placeholder={ __( 'Add profile text...', 'atomic-blocks' ) }
-						keepPlaceholderOnFocus
-						value={ profileContent }
-						formattingControls={ [ 'bold', 'italic', 'strikethrough', 'link' ] }
-						onChange={ ( value ) => setAttributes( { profileContent: value } ) }
-					/>
-
-					<SocialIcons { ...this.props } />
-				</div>
-			</ProfileBox>
-		];
-	}
-}
-
-// Register the block
+/**
+ * Register the block
+ */
 registerBlockType( 'atomic-blocks/ab-profile-box', {
 	title: __( 'AB Profile Box', 'atomic-blocks' ),
 	description: __( 'Add a profile box with bio info and social media links.', 'atomic-blocks' ),
@@ -262,72 +109,17 @@ registerBlockType( 'atomic-blocks/ab-profile-box', {
 		__( 'profile', 'atomic-blocks' ),
 		__( 'atomic', 'atomic-blocks' ),
 	],
-	// Setup the block attributes
+
+	/* Setup the block attributes */
 	attributes: blockAttributes,
 
-	// Render the block components
-	edit: ABAuthorProfileBlock,
+	/* Render the block in the editor. */
+	edit: props => {
+		return <Edit { ...props } />;
+	},
 
-	// Save the block markup
-	save: function( props ) {
-
-		// Setup the attributes
-		const { profileName, profileTitle, profileContent, profileAlignment, profileImgURL, profileImgID, profileFontSize, profileBackgroundColor, profileTextColor, profileLinkColor, twitter, facebook, instagram, pinterest, google, youtube, github, linkedin, email, website, profileAvatarShape } = props.attributes;
-
-		return (
-			// Save the block markup for the front end
-			<ProfileBox { ...props }>
-
-				{ profileImgURL && (
-					<AvatarColumn { ...props }>
-						<div className="ab-profile-image-square">
-							<img
-								className="ab-profile-avatar"
-								src={ profileImgURL }
-								alt="avatar"
-							/>
-						</div>
-					</AvatarColumn>
-				) }
-
-				<div
-					className={ classnames(
-						'ab-profile-column ab-profile-content-wrap'
-					) }
-				>
-					{ profileName && (
-						<RichText.Content
-							tagName="h2"
-							className="ab-profile-name"
-							style={ {
-								color: profileTextColor
-							} }
-							value={ profileName }
-						/>
-					) }
-
-					{ profileTitle && (
-						<RichText.Content
-							tagName="p"
-							className="ab-profile-title"
-							style={ {
-								color: profileTextColor
-							} }
-							value={ profileTitle }
-						/>
-					) }
-
-					{ profileContent && (
-						<RichText.Content
-							tagName="div"
-							className="ab-profile-text"
-							value={ profileContent }
-						/>
-					) }
-
-					<SocialIcons { ...props } />
-				</div>
-			</ProfileBox>
-		);
+	/* Save the block markup. */
+	save: props => {
+		return <Save { ...props } />;
 	},
 } );
